+++ services/storagecache/mgmt/2020-10-01/storagecache/CHANGELOG.md
@@ -1,4 +1,4 @@
-Generated from https://github.com/Azure/azure-rest-api-specs/tree/0b17e6a5e811fd7b122d383b4942441d95e5e8cf//specification/storagecache/resource-manager/readme.md tag: `package-2020-10-01`
+Generated from https://github.com/Azure/azure-rest-api-specs/tree/8240593bde5350e6762015523ccd57cb61e32da5/specification/storagecache/resource-manager/readme.md tag: `package-2020-10-01`
 
 Code generator @microsoft.azure/autorest.go@2.1.171